refactor(game): move shell modules into lib/effects (T30)
Relocate loop, curtain-up, bruff-game, and the observable/ subtree
into packages/game/lib/effects/. After this move lib/ contains only
the five layer directories. loop.ts now reaches state/render/core
via ../...; observable/touch.ts const import drops one extra level
to ../../core/constants.js. Package entry path updated in
package.json and vite.config.lib.ts; vitest coverage exclude paths
updated to the new effects/... locations.
