+++ packages/game/lib/effects/curtain-up.ts
