+++ specs/codebase-compliance/tasks.md
@@ -48,7 +48,7 @@ Each task ends in green `pnpm run ok` for the affected package(s)
 - [x] T27 — Move `packages/game/lib/constants.ts` → `packages/game/lib/core/constants.ts`. Update importers. Seven source importers updated; `vitest.config.ts` coverage `exclude` list also updated to the new path so coverage thresholds remain enforced.
 - [x] T28 — Move `packages/game/lib/create-initial-state.ts`, `update-player.ts`, `update-enemies.ts`, `move-enemy-toward-player.ts` → `packages/game/lib/state/`. Update importers. Test files were moved alongside their sources (eight files total via `git mv`). Within-`state/` cross-imports stayed unchanged (still `./neighbour.js`); imports of `core/` modules became `../core/...`; the only outside importer was `lib/loop.ts`, updated to `./state/...`. `state/.gitkeep` removed.
 - [x] T29 — Move `packages/game/lib/render.ts` → `packages/game/lib/render/render.ts`. Update importers. `render.test.ts` moved alongside its source; the only outside importer was `lib/loop.ts`, updated to `./render/render.js`. `render/.gitkeep` removed.
-- [ ] T30 — Move `packages/game/lib/loop.ts`, `curtain-up.ts`, `bruff-game.ts`, `observable/` → `packages/game/lib/effects/` (preserving the `observable/` subdirectory). Update importers.
+- [x] T30 — Move `packages/game/lib/loop.ts`, `curtain-up.ts`, `bruff-game.ts`, `observable/` → `packages/game/lib/effects/` (preserving the `observable/` subdirectory). Update importers. Two test files moved alongside their sources (`curtain-up.test.ts` and the three `observable/*.test.ts`). `loop.ts` imports of state/render/core became `../...`; `observable/touch.ts` const import became `../../core/constants.js` (one level deeper). Package entry path updated in `package.json` (`main`, `exports.import`) and `vite.config.lib.ts` (`build.lib.entry`); `vitest.config.ts` coverage `exclude` paths updated to the new `effects/...` locations. After this commit `lib/` contains only the five layer directories — no top-level files.
 - [ ] T31 — Add per-layer `no-restricted-imports` rule blocks to `@bruff/eslint-config` covering all five layers. Each block lists only the forbidden upward neighbours.
 - [ ] T32 — Run `pnpm run lint` from repo root; resolve any violation by relocating the offending file, not by widening the rule.
